generateQuerySource関数などをsongTrackRendering.tsに移動
